[Refactor] Move helper to common
