+++ acceptance/openstack/cfw/management/firewall_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/clients"
+	common "github.com/opentelekomcloud/gophertelekomcloud/acceptance/openstack/cfw"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/tools"
 	managementv1 "github.com/opentelekomcloud/gophertelekomcloud/openstack/cfw/v1/management"
 	managementv2 "github.com/opentelekomcloud/gophertelekomcloud/openstack/cfw/v2/management"
@@ -43,7 +44,7 @@ func TestFirewallLifecycle(t *testing.T) {
 	createResp, err := managementv2.Create(clientv2, createOpts)
 	th.AssertNoErr(t, err)
 	th.AssertEquals(t, instanceName, createResp.Data.Name)
-	th.AssertNoErr(t, WaitForJobCompleted(clientv3, 600, 5, createResp.JobID))
+	th.AssertNoErr(t, common.WaitForJobCompleted(clientv3, 600, 5, createResp.JobID))
 	instanceId := createResp.JobID
 	t.Cleanup(func() {
 		_, err = managementv2.Delete(clientv2, instanceId)
